refactor(k8s): simplify manager
